src/requirements.txt
azure-functions>=1.12.0,<2.0.0
azure-storage-blob>=12.14.0,<13.0.0
azure-identity>=1.12.0,<2.0.0
azure-keyvault-secrets>=4.6.0,<5.0.0
openai>=1.1.1,<2.0.0
pytrends>=4.9.0,<5.0.0
requests>=2.28.1,<3.0.0
python-dateutil>=2.8.2,<3.0.0
